requirements.txt
Flask==2.0.1
pyserial==3.5